✨ new Lib, Api response validator and auth header adder (extract from profile server action to commonly using)
