+++ apps/client/src/action/profile/getProfileData.ts
@@ -1,34 +1,18 @@
 "use server"
 
-import { getServerSession } from "next-auth"
-import { authOptions } from "@/app/api/auth/[...nextauth]/authOption"
 import type { ProfileMemberInfoType } from "@/types/profile/profileTypes"
 import type { PromptsType } from "@/types/prompts/promptsType"
+import { isValidResponse } from "@/lib/api/validation"
+import { getAuthHeaders } from "@/lib/api/headers"
 import { profileListData } from "@/dummy/profile/infoAndListDatas"
 
-export async function getProfileMemberInfo(
-	id: string,
-): Promise<ProfileMemberInfoType> {
-	const session = await getServerSession(authOptions)
+export async function getProfileMemberInfo(id: string): Promise<ProfileMemberInfoType> {
+	const headers = await getAuthHeaders()
 	
-	const headers: HeadersInit = {
-		"Content-Type": "application/json",
-	}
-
-	// Add Authorization header only if session exists
-	if (session) {
-		const userObj = session as { user?: { accesstoken?: unknown } }
-		const accessToken = userObj.user?.accesstoken
-
-		if (typeof accessToken === 'string') {
-			headers.Authorization = `Bearer ${accessToken}`
-		}
-	}
-
 	const res = await fetch(`${process.env.API_BASE_URL}/v1/profile/nickname/${id}`, {
 		method: "GET",
 		headers,
-		next: { revalidate: 3600 }, // 1 hour in seconds
+		next: { revalidate: 3600 },
 	})
 
 	if (!res.ok) {
@@ -37,46 +21,20 @@ export async function getProfileMemberInfo(
 
 	const rawData: unknown = await res.json()
 	
-	function isValidResponse(data: unknown): data is { result: ProfileMemberInfoType } {
-		return (
-			typeof data === 'object' && 
-			data !== null && 
-			'result' in data &&
-			typeof (data as { result: unknown }).result === 'object'
-		)
-	}
-
-	if (!isValidResponse(rawData)) {
+	if (!isValidResponse<ProfileMemberInfoType>(rawData)) {
 		throw new Error('Invalid response format')
 	}
 
 	return rawData.result
 }
 
-export async function getProfileMemberInfoByUuid(
-	id: string,
-): Promise<ProfileMemberInfoType> {
-	const session = await getServerSession(authOptions)
+export async function getProfileMemberInfoByUuid(id: string): Promise<ProfileMemberInfoType> {
+	const headers = await getAuthHeaders()
 	
-	const headers: HeadersInit = {
-		"Content-Type": "application/json",
-	}
-
-	// Add Authorization header only if session exists
-	if (session) {
-		const userObj = session as { user?: { accesstoken?: unknown } }
-		const accessToken = userObj.user?.accesstoken
-
-		if (typeof accessToken === 'string') {
-			headers.Authorization = `Bearer ${accessToken}`
-		}
-	}
-
 	const res = await fetch(`${process.env.API_BASE_URL}/v1/profile/uuid/${id}`, {
 		method: "GET",
-		headers,
-		cache: 'force-cache',
-		next: { revalidate: 3600 }, // 1 hour in seconds
+		 headers,
+		next: { revalidate: 3600 },
 	})
 
 	if (!res.ok) {
@@ -85,16 +43,7 @@ export async function getProfileMemberInfoByUuid(
 
 	const rawData: unknown = await res.json()
 	
-	function isValidResponse(data: unknown): data is { result: ProfileMemberInfoType } {
-		return (
-			typeof data === 'object' && 
-			data !== null && 
-			'result' in data &&
-			typeof (data as { result: unknown }).result === 'object'
-		)
-	}
-
-	if (!isValidResponse(rawData)) {
+	if (!isValidResponse<ProfileMemberInfoType>(rawData)) {
 		throw new Error('Invalid response format')
 	}
